requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
yfinance==0.2.41
pytz==2024.1